pyproject.toml
[tool.poetry]
name = "arbitration-bot"
version = "0.0.1"
description = ""
authors = ["Pavlo-Lovchytskyi <pastet1990@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.22.0"
requests = "^2.32.5"
bs4 = "^0.0.2"
lxml = "^6.0.2"
python-dotenv = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
